Enable parallel building while bootstraping qmake.
This reduces build time speciallyon slow-CPU's archs that have more than one
core available.

I have tested this patch without issues, but just for the sake of mental peace
I have also writtrn to Qt's mailing list to ask if this is safe.

I'll later remove this commits if it turns out not to be safe.