Revert "There is no real point in avoiding creating symbols for hese private libs"
This reverts commit 89d190ee7485c037cd421f68ea2bb54514077803.

Actually there is, and quite a lot.

Both libs are private and *only used* by the plugins. In other words,
exclusively accesible for each related plugin which is shipped in the
same binary package.

At this point having shlibs or symbols files referencing them is really
not necessary, moreoever it causes more work for handling symbols files.